requirements-dev.txt
git+https://github.com/botswana-harvard/django-crypto-fields.git@develop#egg=django_crypto_fields
git+https://github.com/botswana-harvard/django-revision.git@develop#egg=django_revision
git+https://github.com/botswana-harvard/edc-action-item.git@develop#egg=edc_action_item
git+https://github.com/botswana-harvard/edc-appointment.git@develop#egg=edc_appointment
git+https://github.com/botswana-harvard/edc-base.git@develop#egg=edc_base
git+https://github.com/botswana-harvard/edc-call-manager.git@develop#egg=edc-call-manager
git+https://github.com/botswana-harvard/edc-constants.git@develop#egg=edc_constants
git+https://github.com/botswana-harvard/edc-consent.git@develop#egg=edc_consent
git+https://github.com/botswana-harvard/edc-data-manager.git@develop#egg=edc-data-manager
git+https://github.com/botswana-harvard/edc-dashboard.git@develop#egg=edc_dashboard
git+https://github.com/botswana-harvard/edc-device.git@develop#egg=edc_device
git+https://github.com/botswana-harvard/edc-fabric.git@develop#egg=edc_fabric
git+https://github.com/botswana-harvard/edc-facility.git@develop#egg=edc_facility
git+https://github.com/botswana-harvard/edc-fieldsets.git@develop#egg=edc_fieldsets
git+https://github.com/botswana-harvard/edc-form-validators.git@develop#egg=edc_form_validators
git+https://github.com/botswana-harvard/edc-identifier.git@develop#egg=edc_identifier
git+https://github.com/botswana-harvard/edc-lab.git@develop#egg=edc_lab
git+https://github.com/botswana-harvard/edc-lab-dashboard.git@develop#egg=edc_lab_dashboard
git+https://github.com/botswana-harvard/edc-label.git@develop#egg=edc_label
git+https://github.com/botswana-harvard/edc-list-data.git@0.1.1#egg=edc-list-data
git+https://github.com/botswana-harvard/edc-locator.git@develop#egg=edc_locator
git+https://github.com/botswana-harvard/edc-metadata.git@develop#egg=edc_metadata
git+https://github.com/botswana-harvard/edc-metadata-rules.git@develop#egg=edc_metadata_rules
git+https://github.com/botswana-harvard/edc-model-admin.git@develop#egg=edc_model_admin
git+https://github.com/botswana-harvard/edc-model-wrapper.git@develop#egg=edc_model_wrapper
git+https://github.com/botswana-harvard/edc-navbar.git@develop#egg=edc_navbar
git+https://github.com/botswana-harvard/edc-prn.git@develop#egg=edc_prn
git+https://github.com/botswana-harvard/edc-protocol.git@develop#egg=edc_protocol
git+https://github.com/botswana-harvard/edc-registration.git@develop#egg=edc_registration
git+https://github.com/botswana-harvard/edc-reference.git@develop#egg=edc_reference
git+https://github.com/botswana-harvard/edc-reports.git@develop#egg=edc_reports
git+https://github.com/botswana-harvard/edc-rest.git@develop#egg=edc_rest
git+https://github.com/botswana-harvard/edc-search.git@develop#egg=edc_search
git+https://github.com/botswana-harvard/edc-sms.git@develop#egg=edc_sms
git+https://github.com/botswana-harvard/edc-subject-dashboard.git@develop#egg=edc_subject_dashboard
git+https://github.com/botswana-harvard/edc-sync@develop#egg=edc_sync
git+https://github.com/botswana-harvard/edc-sync-files@develop#egg=edc_sync_files
git+https://github.com/botswana-harvard/edc-timepoint.git@develop#egg=edc_timepoint
git+https://github.com/botswana-harvard/edc-visit-schedule.git@develop#egg=edc_visit_schedule
git+https://github.com/botswana-harvard/edc-visit-tracking.git@develop#egg=edc_visit_tracking
git+https://github.com/potlako-plus/potlako-dashboard.git@develop#egg=potlako_dashboard
git+https://github.com/potlako-plus/potlako-follow.git@develop#egg=potlako-follow
git+https://github.com/potlako-plus/potlako-metadata-rules.git@develop#egg=potlako_metadata_rules
git+https://github.com/potlako-plus/potlako-prn.git@develop#egg=potlako_prn
git+https://github.com/potlako-plus/potlako-subject.git@develop#egg=potlako_subject
git+https://github.com/potlako-plus/potlako-validations.git@develop#egg=potlako_validations
git+https://github.com/potlako-plus/potlako-visit-schedule.git@develop#egg=potlako_visit_schedule
git+https://github.com/potlako-plus/potlako-prn.git@develop#egg=potlako_prn
git+https://github.com/potlako-plus/potlako-reference.git@develop#egg=potlako_reference
git+https://github.com/potlako-plus/potlako-reports.git@develop#egg=potlako_reports
django-multiselectfield
django-cors-headers
django-crispy-forms
djangorestframework
Fabric3
ipython
mysqlclient
xhtml2pdf
django_q
pycryptodome
xlwt